+++ context/app/static/js/pages/Error/index.ts
@@ -1,0 +1,3 @@
+import ErrorPage from './Error';
+
+export default ErrorPage;