+++ context/app/static/js/shared-styles/sections/SectionFooter/SectionFooter.tsx
@@ -1,0 +1,18 @@
+import React, { Fragment, ReactNode } from 'react';
+
+import { StyledPaper, StyledDivider, StyledTypography } from './style';
+
+function SectionFooter({ items }: { items: { key: string; component: ReactNode }[] }) {
+  return (
+    <StyledPaper>
+      {items.map(({ key, component }, index) => (
+        <Fragment key={key}>
+          {index !== 0 && <StyledDivider orientation="vertical" flexItem />}
+          <StyledTypography variant="caption">{component}</StyledTypography>
+        </Fragment>
+      ))}
+    </StyledPaper>
+  );
+}
+
+export default SectionFooter;